+++ packages/react-mui-audit-log-viewer/src/components/AuditLogViewerComponent/index.tsx
@@ -5,7 +5,7 @@ import { Box } from "@mui/material";
 import { SxProps } from "@mui/system";
 import { DataGridProps, GridColDef, GridSortModel } from "@mui/x-data-grid";
 
-import { Audit } from "../../types";
+import { Audit, FilterOptions } from "../../types";
 import {
   useAuditContext,
   usePagination,
@@ -94,6 +94,8 @@ const AuditLogViewerComponent: FC<ViewerProps> = ({
     setQuery,
     setQueryObj,
     setSort,
+
+    filterOptions,
   } = useAuditContext();
   const { body, bodyWithoutQuery } = useAuditBody(limit, maxResults);
   const pagination = usePagination();
@@ -102,8 +104,8 @@ const AuditLogViewerComponent: FC<ViewerProps> = ({
 
   const schemaColumns = useAuditColumns(schema, fields, fieldTypes);
   const columns = useAuditColumnsWithErrors(schemaColumns, logs);
-  const filterFields = useAuditFilterFields(schema);
-  const conditionalOptions = useAuditConditionalOptions(schema);
+  const filterFields = useAuditFilterFields(schema, filterOptions);
+  const conditionalOptions = useAuditConditionalOptions(schema, filterOptions);
 
   const hasMountedRef = useRef<string | undefined>(undefined);