+++ packages/react-mui-shared/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.0.4] - 2025-02-25
+
+### Added
+
+- Updated TimeRangeSelect to support children prop and quickTimeRanges prop to override the default quick time range options of (1 day, 7 days, and 30 days)
+- Exporting RelativeDateRangeField component relevant and utility functions
+
 ## [1.0.3] - 2024-12-18
 
 ### Added